+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWE79MQ1Y4F4HBKP6VPCKYF8"
+        "01KWECGYV8S58P9ZGAGXZB44Q4"
     )
     r"""The unique identifier of the data source"""