+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWE79MQ5F9R2QNSZ4KTYN0BR"
+        "01KWECGYVCBMW5T2D6PPQ7T1HS"
     )
     r"""The unique identifier of the data source"""